+++ skbio/io/format/gff3.py
@@ -1,4 +1,4 @@
-r'''
+'''
 GFF3 format (:mod:`skbio.io.format.gff3`)
 =========================================
 
@@ -15,153 +15,90 @@ Format Support
 +------+------+---------------------------------------------------------------+
 |Reader|Writer|                          Object Class                         |
 +======+======+===============================================================+
-|Yes   |Yes   |:mod:`skbio.metadata.IntervalMetadata`                         |
+|Yes   |Yes   |:mod:`skbio.sequence.Sequence`                                 |
 +------+------+---------------------------------------------------------------+
 |Yes   |Yes   |:mod:`skbio.sequence.DNA`                                      |
 +------+------+---------------------------------------------------------------+
-|Yes   |Yes   |:mod:`skbio.sequence.Sequence`                                 |
+|Yes   |Yes   |:mod:`skbio.metadata.IntervalMetadata`                         |
 +------+------+---------------------------------------------------------------+
-|Yes   |Yes   |generator of :mod:`skbio.metadata.IntervalMetadata`            |
+|Yes   |Yes   |generator of tuple (seq_id of str type,                        |
+|      |      |:mod:`skbio.metadata.IntervalMetadata`)                        |
 +------+------+---------------------------------------------------------------+
 
 Format Specification
 --------------------
+**State: Experimental as of 0.5.1.**
+
 The first line of the file is a comment that identifies the format and
-version.  This is followed by a series of data lines, each one of
-which corresponds to an annotation.  The 9 columns of the annotation
-section are as follows:
-
-+-------+------------------------------------------+
-|Column | Description                              |
-+=======+==========================================+
-|SEQID  | ID of the landmark used                  |
-+-------+------------------------------------------+
-|SOURCE | algorithm used to generate this feature  |
-+-------+------------------------------------------+
-|TYPE   | type of the feature                      |
-+-------+------------------------------------------+
-|START  | start of the feature                     |
-+-------+------------------------------------------+
-|END    | end of the feature                       |
-+-------+------------------------------------------+
-|SCORE  | floating point score                     |
-+-------+------------------------------------------+
-|STRAND | The strand of the feature (+/-/./?)      |
-+-------+------------------------------------------+
-|PHASE  | only for TYPE="CDS"                      |
-+-------+------------------------------------------+
-|ATTR   | feature attributes                       |
-+-------+------------------------------------------+
-
+version. This is followed by a series of data lines. Each data line
+corresponds to an annotation and consists of 9 columns: SEQID, SOURCE,
+TYPE, START, END, SCORE, STRAND, PHASE, and ATTR.
 
 Column 9 (ATTR) is list of feature attributes in the format
-tag=value. Multiple tag=value pairs are separated by
-semicolons. Multiple values of the same tag are indicated by
-separating the values with the comma ",". The following tags have
-predefined meanings:
+"tag=value". Multiple "tag=value" pairs are delimited by
+semicolons. Multiple values of the same tag are separated with the
+comma ",". The following tags have predefined meanings: ID, Name,
+Alias, Parent, Target, Gap, Derives_from, Note, Dbxref, Ontology_term,
+and Is_circular.
 
-* ID. Indicates the unique identifier of the feature. IDs must be
-  unique within the scope of the GFF file.
-
-* Name. Display name for the feature. This is the name to be displayed
-  to the user. Unlike IDs, there is no requirement that the Name be
-  unique within the file.
-
-* Alias. A secondary name for the feature. It is suggested that this
-  tag be used whenever a secondary identifier for the feature is
-  needed, such as locus names and accession numbers. Unlike ID, there
-  is no requirement that Alias be unique within the file.
-
-* Parent. Indicates the parent of the feature. A parent ID can be used
-  to group exons into transcripts, transcripts into genes and so
-  forth. A feature may have multiple parents. Parent can *only* be
-  used to indicate a partof relationship.
-
-* Target. Indicates the target of a nucleotide-to-nucleotide or
-  protein-to-nucleotide alignment. The format of the value is
-  "target_id start end [strand]", where strand is optional and may be
-  "+" or "-". If the target_id contains spaces, they must be escaped
-  as hex escape %20.
-
-* Gap. The alignment of the feature to the target if the two are not
-  collinear (e.g. contain gaps). The alignment format is taken from
-  the CIGAR format described in the Exonerate documentation.
-
-* Derives_from. Used to disambiguate the relationship between one
-  feature and another when the relationship is a temporal one rather
-  than a purely structural "part of" one. This is needed for
-  polycistronic genes.
-
-* Note. A free text note.
-
-* Dbxref. A database cross reference. See the GFF3 specification for
-  more information.
-
-* Ontology_term. A cross reference to an ontology term.
-
-* Is_circular. A flag to indicate whether a feature is circular.
-
-The columns and attributes are read in as the vocabulary defined in
-genbank parsers (:mod:`skbio.io.format.genbank`).
+The meaning and format of these columns and attributes are explained
+detail in the format specification [1]_. And they are read in as the
+vocabulary defined in GenBank parser (:mod:`skbio.io.format.genbank`).
 
 Format Parameters
 -----------------
 
 Reader-specific Parameters
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
-``interval_metadata_dict`` is a ``dict`` parameter required for
-``IntervalMetadata`` generator. Its key is the ``str`` of sequence ID
-and its value is the ``IntervalMetadata`` object for the sequence. The
-generator will read the GFF3 file and add the annotation of sequences
-to the corresponding ``IntervalMetadata`` objects. If the annotations of
-some sequences in the GFF3 doesn't match any sequence IDs in the input
-dict of ``interval_metadata_dict``, those annotations will be skipped
-and not parsed.
+Reader of ``IntervalMetadata`` generator requires a parameter
+``lengths``. It is an iterable of int and specifies the upper bounds
+of the ``IntervalMetadata`` objects the reader yields.
 
-``interval_metadata`` is an ``IntervalMetadata`` object required for
-``IntervalMetadata`` GFF3 reader. It reads the GFF3 file and add the
-annotations to the input ``interval_metadata``.
+``IntervalMetadata`` GFF3 reader requires 2 parameters: ``seq_id`` of
+str and ``length`` of int. It reads the annotation with the specified
+sequence ID from the GFF3 file into an ``IntervalMetadata`` object
+with upper bound of ``length``.
 
-``rec_num`` is an ``int` parameter used with the ``IntervalMetadata``,
-``DNA``, and ``Sequence`` GFF3 readers. It specifies which GFF3 record
-to read from a GFF3 file with annotations of multiple sequences in it.
+``DNA`` and ``Sequence`` GFF3 readers require ``seq_num`` of int as
+parameter. It specifies which GFF3 record to read from a GFF3 file
+with annotations of multiple sequences in it.
 
 Writer-specific Parameters
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
-``skip`` is a boolean parameter used by all the GFF3 writers. It
+``skip_subregion`` is a boolean parameter used by all the GFF3 writers. It
 specifies whether you would like to write each non-contiguous
-sub-regions for a feature annotation. For example, if there is
+sub-region for a feature annotation. For example, if there is
 interval feature for a gene with two exons in an ``IntervalMetadata``
-object, it will write one line into the GFF3 file when ``skip`` is
+object, it will write one line into the GFF3 file when ``skip_subregion`` is
 ``True`` and will write 3 lines (one for the gene and one for each
-exon, respectively) when ``skip`` is ``False``. Default is ``True``.
+exon, respectively) when ``skip_subregion`` is ``False``. Default is ``True``.
 
-``seq_id`` is a ``str`` parameter required to write
-``IntervalMetadata`` to GFF3. It specify the sequence ID (column 1 in
-GFF3 file) that the annotation belong to.
-
-``seq_ids`` is an iterable parameter required for ``IntervalMetadata``
-generator. It specifies each sequence IDs for ``IntervalMetadata``
-objects.
+In addition, ``IntervalMetadata`` GFF3 writer needs a parameter of
+``seq_id``. It specify the sequence ID (column 1 in GFF3 file) that
+the annotation belong to.
 
 Examples
 --------
 
+Let's create a file stream with following data in GFF3 format:
+
 >>> gff_str = """
-... ##gff-version\t3.2.1
-... ##sequence-region\tctg123\t1\t1497228
-... ctg123\t.\tgene\t1000\t9000\t.\t+\t0\tID=gene00001;Name=EDEN
-... ctg123\t.\tTF_binding_site\t1000\t1012\t.\t+\t.\tParent=gene00001
-... ctg123\t.\tmRNA\t1050\t9000\t.\t+\t.\tID=mRNA00001;Parent=gene00001
+... ##gff-version\\t3.2.1
+... ##sequence-region\\tctg123\\t1\\t1497228
+... ctg123\\t.\\tgene\\t1000\\t9000\\t.\\t+\\t0\\tID=gene00001;Name=EDEN
+... ctg123\\t.\\tTF_binding_site\\t1000\\t1012\\t.\\t+\\t.\\tParent=gene00001
+... ctg123\\t.\\tmRNA\\t1050\\t9000\\t.\\t+\\t.\\tID=mRNA00001;Parent=gene00001
 ... """
 >>> import io
 >>> from skbio.metadata import IntervalMetadata
 >>> from skbio.io import read
->>> im = IntervalMetadata(20000)
 >>> gff = io.StringIO(gff_str)
->>> im_return = read(gff, format='gff3', into=IntervalMetadata,
-...                  interval_metadata=im)
->>> im_return   # doctest: +SKIP
+
+We can read it into
+
+>>> im = read(gff, format='gff3', into=IntervalMetadata,
+...           seq_id='ctg123', length=10000)
+>>> im   # doctest: +SKIP
 3 interval features
 -------------------
 Interval(interval_metadata=<4601272528>, bounds=[(999, 9000)], fuzzy=\
@@ -173,12 +110,12 @@ Interval(interval_metadata=<4601272528>, bounds=[(999, 1012)], fuzzy=\
 Interval(interval_metadata=<4601272528>, bounds=[(1049, 9000)], fuzzy=\
 [(False, False)], metadata={'source': '.', 'type': 'mRNA', 'strand': '+', \
 'score': '.', 'ID': 'mRNA00001', 'Parent': 'gene00001'})
->>> im == im_return
-True
+
 
 References
 ----------
-.. [1] https://github.com/The-Sequence-Ontology/Specifications/blob/master/gff3.md  # noqa
+.. [1] https://github.com/The-Sequence-Ontology/\
+Specifications/blob/master/gff3.md
 
 '''
 
@@ -225,93 +162,87 @@ def _gff3_sniffer(fh):
 
 
 @gff3.reader(None)
-def _gff3_to_generator(fh, interval_metadata_dict):
+def _gff3_to_generator(fh, lengths):
     '''Parse the GFF3 into the existing IntervalMetadata
-
-    Note that if the seq ID does not exist in the input dict of
-    ``interval_metadata_dict``, the record will be skipped and not
-    parsed.
 
     Parameters
     ----------
     fh : file
         file handler
-    interval_metadata_dict : dict
-        key is seq ID and value is the IntervalMetadata for the seq.
 
     Yields
     ------
-    IntervalMetadata
+    tuple
+        str of seq id, IntervalMetadata
     '''
-    for seq_id, lines in _yield_record(fh):
-        if seq_id in interval_metadata_dict:
-            imd = interval_metadata_dict[seq_id]
-            yield _parse_record(lines, imd)
+    for (seq_id, lines), length in zip(_yield_record(fh), lengths):
+        yield seq_id, _parse_record(lines, length)
 
 
 @gff3.writer(None)
-def _generator_to_gff3(obj, fh, seq_ids, skip=True):
+def _generator_to_gff3(obj, fh, skip_subregion=True):
     '''Write list of IntervalMetadata into file.
 
     Parameters
     ----------
-    obj : Iterable of IntervalMetadata
+    obj : Iterable of (seq_id, IntervalMetadata)
     fh : file handler
-    seq_ids : Iterable of seq id (str)
     '''
-    for obj_i, seq_id in zip(obj, seq_ids):
-        _serialize_interval_metadata(obj_i, seq_id, fh, skip)
+    for seq_id, obj_i in obj:
+        _serialize_interval_metadata(obj_i, seq_id, fh, skip_subregion)
 
 
 @gff3.reader(Sequence)
-def _gff3_to_sequence(fh, rec_num=1):
+def _gff3_to_sequence(fh, seq_num=1):
     ''''''
-    seq_id, lines = _get_nth_record(_yield_record(fh), rec_num)
+    seq_id, lines = _get_nth_record(_yield_record(fh), seq_num)
     # you can't read directly from fh because in registry.py line 543
     # file.tell() will fail "telling position disabled by next() call".
     stream = StringIO(fh.read())
-    seq = read(stream, format='fasta', into=Sequence, seq_num=rec_num)
-    _parse_record(lines, interval_metadata=seq.interval_metadata)
+    seq = read(stream, format='fasta', into=Sequence, seq_num=seq_num)
+    seq.interval_metadata=_parse_record(lines, len(seq))
     return seq
 
 
 @gff3.writer(Sequence)
-def _sequence_to_gff3(obj, fh, skip=True):
-    _serialize_seq(obj, fh, skip)
+def _sequence_to_gff3(obj, fh, skip_subregion=True):
+    _serialize_seq(obj, fh, skip_subregion)
 
 
 @gff3.reader(DNA)
-def _gff3_to_dna(fh, rec_num=1):
+def _gff3_to_dna(fh, seq_num=1):
     ''''''
-    seq_id, lines = _get_nth_record(_yield_record(fh), rec_num)
+    seq_id, lines = _get_nth_record(_yield_record(fh), seq_num)
     stream = StringIO(fh.read())
-    seq = read(stream, format='fasta', into=DNA, seq_num=rec_num)
-    _parse_record(lines, interval_metadata=seq.interval_metadata)
+    seq = read(stream, format='fasta', into=DNA, seq_num=seq_num)
+    seq.interval_metadata = _parse_record(lines, len(seq))
     return seq
 
 
 @gff3.writer(DNA)
-def _dna_to_gff3(obj, fh, skip=True):
-    _serialize_seq(obj, fh, skip)
+def _dna_to_gff3(obj, fh, skip_subregion=True):
+    _serialize_seq(obj, fh, skip_subregion)
 
 
 @gff3.reader(IntervalMetadata)
-def _gff3_to_interval_metadata(fh, interval_metadata, rec_num=1):
+def _gff3_to_interval_metadata(fh, seq_id, length):
     '''Read a GFF3 record into the specified interval metadata.
 
     Parameters
     ----------
     fh : file handler
-    interval_metadata : IntervalMetadata
-    rec_num : int
-        which record to read in.
+    length : int
+        seq length
     '''
-    seq_id, lines = _get_nth_record(_yield_record(fh), rec_num)
-    return _parse_record(lines, interval_metadata=interval_metadata)
+    for sid, lines in _yield_record(fh):
+        if sid == seq_id:
+            return _parse_record(lines, length)
+    # return an empty instead of None
+    return IntervalMetadata(length)
 
 
 @gff3.writer(IntervalMetadata)
-def _interval_metadata_to_gff3(obj, fh, seq_id, skip=True):
+def _interval_metadata_to_gff3(obj, fh, seq_id, skip_subregion=True):
     '''
     Parameters
     ----------
@@ -319,11 +250,11 @@ def _interval_metadata_to_gff3(obj, fh, seq_id, skip=True):
     seq_id : str
         ID for column 1 in the GFF3 file.
     '''
-    _serialize_interval_metadata(obj, seq_id, fh, skip=True)
+    _serialize_interval_metadata(obj, seq_id, fh, skip_subregion=True)
 
 
 def _yield_record(fh):
-    '''Yield lines that belong to the same sequence.'''
+    '''Yield (seq_id, lines) that belong to the same sequence.'''
     lines = []
     current = False
     for line in _line_generator(fh, skip_blanks=True, strip=True):
@@ -346,8 +277,9 @@ def _yield_record(fh):
     yield current, lines
 
 
-def _parse_record(lines, interval_metadata):
+def _parse_record(lines, length):
     '''Parse the lines into a IntervalMetadata object.'''
+    interval_metadata = IntervalMetadata(length)
     for line in lines:
         columns = line.split('\t')
         # there should be 9 columns
@@ -393,17 +325,17 @@ def _parse_attr(s):
 
 
 def _serialize_interval_metadata(
-        interval_metadata, seq_id, fh, skip=True):
+        interval_metadata, seq_id, fh, skip_subregion=True):
     '''Serialize an IntervalMetadata to GFF3.
 
     Parameters
     ----------
     interval_metadata : IntervalMetadata
-    skip : bool
+    skip_subregion : bool
         whether to skip outputting each sub region as a line in GFF3.
     '''
     # write file header
-    print('##gff-version 3', file=fh)
+    fh.write('##gff-version 3\n')
 
     column_keys = ['source', 'type', 'score', 'strand', 'phase']
     voca_change = _vocabulary_change('gff3', False)
@@ -444,7 +376,7 @@ def _serialize_interval_metadata(
 
         # if there are multiple regions for this feature,
         # output each region as a standalone line in GFF3.
-        if len(bd) > 1 and skip is False:
+        if len(bd) > 1 and skip_subregion is False:
             for start, end in bd:
                 # if this is a gene, then each sub region should be an exon
                 if columns[2] == 'gene':
@@ -461,9 +393,9 @@ def _serialize_interval_metadata(
                 print('\t'.join(columns), file=fh)
 
 
-def _serialize_seq(seq, fh, skip=True):
+def _serialize_seq(seq, fh, skip_subregion=True):
     '''Serialize a sequence to GFF3.'''
     _serialize_interval_metadata(
-        seq.interval_metadata, seq.metadata['id'], fh, skip)
+        seq.interval_metadata, seq.metadata['id'], fh, skip_subregion)
     print('##FASTA', file=fh)
     write(seq, into=fh, format='fasta')